+++ attestation_card.php
@@ -411,7 +411,8 @@ if ($action == 'create' && empty($typeCode)) {
 	print '<div class="fichecenter">';
 	print '<div class="fichehalfleft">';
 	$uploadDir = powerplantpvAttestationGetDocumentUploadDir($object);
-	$formfile->showdocuments(powerplantpvAttestationGetDocumentModulePart().':PowerPlantPVAttestation', powerplantpvAttestationGetDocumentRelativePath($object), $uploadDir, $_SERVER['PHP_SELF'].'?id='.(int) $object->id, $permissiontoadd, $permissiontodelete, $object->model_pdf, 1, 0, 0, 28, 0, '', '', '', $langs->defaultlang, '', $object);
+	powerplantpvAttestationPrintDocumentGenerationForm($object, $_SERVER['PHP_SELF'].'?id='.(int) $object->id, $permissiontoadd, $object->model_pdf, 0, '', $langs->defaultlang);
+	$formfile->showdocuments(powerplantpvAttestationGetDocumentModulePart(), powerplantpvAttestationGetDocumentRelativePath($object), $uploadDir, $_SERVER['PHP_SELF'].'?id='.(int) $object->id, 0, $permissiontodelete, $object->model_pdf, 1, 0, 0, 28, 0, '', 'none', '', $langs->defaultlang, '', $object);
 	if (method_exists($form, 'showLinkedObjectBlock')) {
 		$form->showLinkedObjectBlock($object);
 	}